Uses speed and distance to calculate the interval with the car in front.
Uses an EMA filter to reduce jitter. The jitter is due to the fact that the
speed is not constant, so the distance can fluctuate a lot when braking and
accelerating. The smaller the smooth period, the more sensitive the system is an
the less stable; the larger the period, the less sensitive it is to change but
more stable.